+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Feature: Fetch ERC-20 details
 - Feature: Gas estimator
 - Feature: Monitor success/fail execution of transactions in parallel
+- Feature: Analyse Uniswap trades to get the price and fees paid
 
 # 0.1